+++ packages/blade/src/components/Accordion/AccordionItemHeader.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { useAccordion, useAccordionItemIndex } from './AccordionContext';
 import { componentIds } from './componentIds';
 import type { BaseHeaderProps } from '~components/BaseHeaderFooter/BaseHeader';
@@ -23,8 +24,24 @@ const _AccordionItemHeader = ({
   'title' | 'subtitle' | 'leading' | 'children' | 'trailing' | 'titleSuffix'
 > &
   DataAnalyticsAttribute): React.ReactElement => {
-  const { size, showNumberPrefix } = useAccordion();
+  const { size, showNumberPrefix, expandedIndex } = useAccordion();
   const { index, isDisabled } = useAccordionItemIndex();
+
+  const isLeadingNumberOrIcon = React.useMemo(() => {
+    // Check if leading is a number
+    if (showNumberPrefix && typeof index === 'number') return true;
+
+    // Check if leading is an Icon component (name ends with "Icon")
+    if (
+      leading &&
+      React.isValidElement(leading) &&
+      typeof leading.type === 'function' &&
+      leading.type.name?.endsWith('Icon')
+    )
+      return true;
+
+    return false;
+  }, [leading, showNumberPrefix, index]);
 
   return (
     <BaseBox {...metaAttribute({ name: MetaConstants.AccordionItemHeader })} flex="1">
@@ -46,10 +63,15 @@ const _AccordionItemHeader = ({
         isDisabled={isDisabled}
         showBackButton={false}
         showCloseButton={false}
-        showDivider={false}
+        showDivider={expandedIndex === index}
         paddingX="spacing.5"
         marginY="spacing.5"
         size={size}
+        alignItems={!subtitle && !isLeadingNumberOrIcon ? 'center' : 'flex-start'}
+        dividerProps={{
+          thickness: 'thinner',
+          marginX: 'spacing.5',
+        }}
         trailingInteractionElement={
           <CollapsibleChevronIcon
             color={isDisabled ? 'interactive.icon.gray.disabled' : 'interactive.icon.gray.muted'}