+++ packages/blade/src/components/Button/BaseButton/BaseButton.tsx
@@ -77,7 +77,15 @@ type BaseButtonCommonProps = {
   isLoading?: boolean;
   accessibilityProps?: Partial<AccessibilityProps>;
   variant?: 'primary' | 'secondary' | 'tertiary';
-  color?: 'primary' | 'white' | 'positive' | 'negative' | 'notice' | 'information' | 'neutral';
+  color?:
+    | 'primary'
+    | 'white'
+    | 'positive'
+    | 'negative'
+    | 'notice'
+    | 'information'
+    | 'neutral'
+    | 'transparent';
 } & TestID &
   StyledPropsBlade &
   BladeCommonEvents;
@@ -136,6 +144,15 @@ export const getBackgroundColorToken = ({
     return tokens.white[variant][_state];
   }
 
+  if (color === 'transparent') {
+    if (variant !== 'tertiary') {
+      throw new Error(
+        `Transparent color can only be used with tertiary variant but received "${variant}"`,
+      );
+    }
+    return tokens.transparent.tertiary[_state];
+  }
+
   if (color && color !== 'primary') {
     if (variant === 'tertiary') {
       throw new Error(
@@ -161,6 +178,15 @@ export const getTextColorToken = ({
 
   if (color === 'white') {
     return tokens.white[variant][_state];
+  }
+
+  if (color === 'transparent') {
+    if (variant !== 'tertiary') {
+      throw new Error(
+        `Transparent color can only be used with tertiary variant but received "${variant}"`,
+      );
+    }
+    return tokens.transparent.tertiary[_state];
   }
 
   if (color && color !== 'primary') {
@@ -194,10 +220,15 @@ const getProps = ({
   variant: NonNullable<BaseButtonProps['variant']>;
   color: BaseButtonProps['color'];
 }): BaseButtonStyleProps => {
-  if (variant === 'tertiary' && color !== 'primary' && color !== 'white') {
+  if (
+    variant === 'tertiary' &&
+    color !== 'primary' &&
+    color !== 'white' &&
+    color !== 'transparent'
+  ) {
     throwBladeError({
       moduleName: 'BaseButton',
-      message: `Tertiary variant can only be used with color: "primary" or "white" but received "${color}"`,
+      message: `Tertiary variant can only be used with color: "primary" or "white" or "transparent" but received "${color}"`,
     });
   }
 
@@ -250,6 +281,15 @@ const getProps = ({
     hoverBorderColor: getIn(
       theme.colors,
       getBackgroundColorToken({ property: 'border', variant, color, state: 'hover' }),
+    ),
+    hoverIconColor: getIn(
+      theme.colors,
+      getTextColorToken({
+        property: 'icon',
+        variant,
+        color,
+        state: 'hover',
+      }),
     ),
     focusBackgroundColor: getIn(
       theme.colors,
@@ -382,6 +422,7 @@ const _BaseButton: React.ForwardRefRenderFunction<BladeElementRef, BaseButtonPro
     fontSize,
     hoverBorderColor,
     hoverBackgroundColor,
+    hoverIconColor,
     iconColor,
     iconSize,
     iconPadding,
@@ -455,6 +496,7 @@ const _BaseButton: React.ForwardRefRenderFunction<BladeElementRef, BaseButtonPro
         }),
       }}
       variant={variant}
+      color={color}
       isLoading={isLoading}
       disabled={disabled}
       defaultBorderColor={defaultBorderColor}
@@ -499,6 +541,11 @@ const _BaseButton: React.ForwardRefRenderFunction<BladeElementRef, BaseButtonPro
       height={height}
       width={width}
       isPressed={isPressed}
+      hoverIconColor={
+        isDisabled
+          ? getIn(theme.colors, iconColor as DotNotationToken<Theme['colors']>)
+          : hoverIconColor
+      }
       onMouseDown={(event: React.MouseEvent) => {
         handlePointerPressedIn();
         onMouseDown?.(event);
@@ -527,7 +574,11 @@ const _BaseButton: React.ForwardRefRenderFunction<BladeElementRef, BaseButtonPro
             right="0px"
             zIndex={1}
           >
-            <BaseSpinner accessibilityLabel="Loading" size={spinnerSize} color={color} />
+            <BaseSpinner
+              accessibilityLabel="Loading"
+              size={spinnerSize}
+              color={color === 'transparent' ? 'primary' : color}
+            />
           </BaseBox>
         ) : null}
         <ButtonContent