+++ packages/blade/src/components/CreationView/CreationView.stories.tsx
@@ -67,6 +67,7 @@ const Page = (): React.ReactElement => {
       componentDescription="Creation View is a pattern that is used in creation flows"
       apiDecisionLink={null}
       figmaURL="https://www.figma.com/design/ZZ2dpcIAsPCEGPwQ2UdgL1/Blade-Cheatsheet?node-id=949-178337&m=dev"
+      codeUrl="https://github.com/razorpay/blade/blob/master/packages/blade/src/components/CreationView/CreationView.stories.tsx"
     >
       <Heading size="large">Usage</Heading>
       <Sandbox showConsole>