+++ packages/blade/src/components/InfoGroup/index.ts
@@ -1,0 +1,2 @@
+export { InfoGroup, InfoItem, InfoItemKey, InfoItemValue } from './InfoGroup';
+export * from './types';