+++ packages/blade/src/components/Input/BaseInput/AnimatedBaseInputWrapper.web.tsx
@@ -152,6 +152,7 @@ to {
         }
       }}
       isTableInputCell={rest.isTableInputCell}
+      className="__blade-base-input-wrapper"
     />
   );
 };