+++ packages/blade/src/components/ListView/types.ts
@@ -8,7 +8,7 @@ type ListViewCommonProps = {
 type ListViewProps = ListViewCommonProps & TestID & DataAnalyticsAttribute;
 
 type ListViewFilterProps = {
-  children: React.ReactNode;
+  children?: React.ReactNode;
   /**
    * Quick Filters Component
    */
@@ -48,8 +48,7 @@ type ListViewFilterProps = {
    */
   selectedFiltersCount?: number;
 } & TestID &
-  DataAnalyticsAttribute &
-  ListViewCommonProps;
+  DataAnalyticsAttribute;
 
 type ListViewSelectedFiltersType = {
   [key: string]: string[] | string | number[];